#4 feat: Add new MCQ exercise
- Introduced `fetch_node_with_metadata` and `search_nodes` functions in the Rust API.
- Implemented `get_word_instance_context` to gather context for MCQ exercises related to word instances.
- Enhanced the `SqliteRepository` to support new queries for fetching node metadata and word instance context.
- Updated `SseDecode` and `SseEncode` implementations for new data structures including `NodeData` and `WordInstanceContext`.
- Modified the exercise handling to support new types and serialization formats.

+++ lib/widgets/debug_panel.dart
@@ -149,7 +149,9 @@ class _DebugPanelDialogState extends ConsumerState<_DebugPanelDialog> {
   }
 
   static Widget _buildDueItemCard(DueItem item) {
-    final dueTime = DateTime.fromMillisecondsSinceEpoch(item.state.dueAt);
+    final dueTime = DateTime.fromMillisecondsSinceEpoch(
+      item.state.dueAt.toInt(),
+    );
     final now = DateTime.now();
     final isOverdue = dueTime.isBefore(now);
     final timeStr = DateFormat('MMM dd HH:mm').format(dueTime);
